+++ CHANGES.rst
@@ -1,11 +1,26 @@
 Products.MeetingCharleroi Changelog
 ===================================
 
-4.1.8 (unreleased)
-------------------
+
+4.1.10 (unreleased)
+-------------------
 
 - Nothing changed yet.
 
+4.1.9 (2021-12-03)
+------------------
+
+- Keep pollType when item is duplicated to other MeetingConfig.
+  [aduchene]
+- Fixed `test_ValidateCategoryIfCollegeItemToSendToCouncil` due to change in
+  `MeetingItem.validate_category`, the given category must exists.
+  [gbastien]
+
+4.1.8 (2021-07-26)
+------------------
+
+- Added a default classifier when an item is sent from one MeetingConfig to another.
+  [aduchene]
 
 4.1.7 (2020-10-30)
 ------------------